Makefile: add shushtar prefix to commit version
The version string on boot up displays the lnd version hard coded in
lnd/build/version.go but the commit of shushtar (currently
v0.1.0-alpa.rcX). This could be confusing to users, why the second
version is completely different. I added the prefix "shushtar-" to the
commit to make it clear it is referencing the shushtar repo.

We also fix the Makefile to properly compile in the commit hash.