terminal+subserver: validate subserver macaroons
When running in the same process, lnd will not know what to do with a
macaroon that is meant for a subserver. With this commit we instruct the
interceptor in the RPC proxy to delegate the validation of any macaroon
unknown to lnd back to the terminal instance which then passes it on to
the subserver it was meant for.