types: add generated code from proto compilation to git
To remove the need for installing protoc for building the source from
scratch, we check in the generated JavaScript and TypeScript code that
is compiled from the *.proto files.
We will make sure the protos and generated files are in sync with the
help of a GitHub Action.
By adding a .gitattributes file we also make sure the committed
generated JS code does not count towards GitHub's language statistics of
the repo.