doc: unify shell code block formatting
The code blocks aren't displayed correctly in GitBook when using the
"shell script" markup hint. We unify all code blocks to use the same
"shell" hint. While doing so, we also use an explicit unicode character
to represent the shell. This should make it more clear that the
character isn't part of the command itself.