doc+config: move network config to top-level struct
Since we know that only the Bitcoin network is used and that all
components must operate on the same network, we can make the
configuration a lot easier for the user by allowing them to only specify
a single option for the network. We'll then go ahead and set that in all
the necessary sub structs.