terminal: register integrated daemon gRPC server to proxy
It turns out we never implemented the proxying of the daemon gRPC
requests as drawn in the ASCII diagram. We used to hook up the daemon
gRPC servers to the instance we got back from lnd. The requests needed
to be forwarded internally to lnd's gRPC listener even though they'd be
processed by the external subservers later.
We can short-circuit that whole redirection by simply hooking the daemon
gRPC servers directly into the proxy's gRPC server instance.
This also fixes the issue that Faraday, Loop and Pool requests wouldn't
work anymore in integrated mode after the recent changes in the proxy
logic.