firewalldb: add Action CRUD logic
This commit adds a new Action type to the firewalldb package along with
Serialization and Deserialization methods for it. It also adds
ListAction and ListSessionAction methods to the DB which can be used to
paginate through the actions. Various interfaces are also added that
will be used by `rules` to access certain RuleActions.