chore: prevent version updates by dependabot
It turns out that the `dependabot.yml` config file only applies to *version* updates, not *security* updates. When I added this file in #396, the bot decided to open PRs for some of our production dependencies, even if there isn't a security issue with those packages. I don't like the automated PRs because it turns into a bunch of spam in the repo. I'm deleting this file to turn off the dependabot version updates.