multi: add withAllSubServers param to perms.NewManager
In this commit, a new withAllSubServers option is added to the
perms.NewManager function. If its value is true, then all LND's
subserver permissions will be added to the active permissions set
regardless of whether LND was actually compiled with the necessary build
tags. This is useful for Lit clients who perhaps can never know which
sub-servers LND was compiled with and so want to assume that it was
compiled with all.