app/build: add .gitkeep to make sure directory exists
Even though the content of the app/build directory should be ignored by
git, we need it to exist in order for the Golang embedded file system
compilation to succeed on an empty project. This is required to fix
compilation issues when lightning-terminal is used as a module
dependency in another project, where the frontend build process can't be
triggered.