config: run taproot assets subserver by default
By default the taproot assets subserver will start in integrated mode
like the rest of subservers. An execption is if we are running in
"mainnet" where we want it disabled because it is not yet supported.

Users won't be able to start the sub-server in "mainnet" even if they
set the `--taproot-assets-mode=` flag.