mod: Update taproot-assets dependency
Update the taproot-assets dependency to a version of the main branch
that includes a commit which contains a fix that removes the js build
tag requirement when using the taprpc packages' JSON callback functions.

Also replace the google.golang.org/protobuf dependency to forked version
which is required for the taproot-assets dependency to work properly.