session: add session linking fields to Session
Add the new `PrevLocalPub` and `GroupID` key to the `Session` struct and
ensure that the new fields are properly serialised and deserialised. The
PrevLocalPub is an optional field that can be used to link to a prior
active session by setting the field equal to the LocalPublicKey of a
prior session. The `GroupID` is the ID of the first Session in a set of
linked sessions.