litrpc: add session link field to AddAutopilotSessionRequest
Update the Autopilot service AddAutopilotSessionRequest message to
include a new `prev_local_pub` field. Users may set this field to the
local public key of a previously active Session in order to link the new
session to the old one. The Session message is also updated with new
`prev_session_local_pub` field which will be populated by the local pub
key of a linked session. A new group_id field is also added which is the
ID of the group of linked sessions that the session belongs to.