terminal: dont block indefinitely on macaroon channel
In this commit, we ensure that when we wait on the macaroon channel for
the macaroon from the integrated LND, that we also listen to other exit
signals such as the interceptor, lndQuit chanel and the error channel.
This is to prevent shutdown from blocking forever in the case where
there was an issue preventing LND from sending the initial macaroon.