multi: bump lnd, loop & taproot-assets versions
Bump lnd to `v0.17.1-beta`, loop to `v0.26.5-beta` and taproot-assets to
`v0.3.1-alpha`.

With commit `76862ed` in lnd v0.17.1-beta, lnd will return an
"invalid ID" error for any calls with an unknown macaroon version.
Therefore our itests that uses a dummy macaroon was also updated to
expect this error.

With tapd `v0.3.1-alpha`, the universerpc Info rpc response is changed,
which removes the `num_assets` field in the response.
This commit therefore also updates the `taprpc-whitelist` itest
successPattern to match the new response.