status: add status manager isReadyOverride func
isReadyOverride is a callback that, when set and only if `running` is
not yet true, will be used to determine if a system is ready for a call.
We will pass the request URI to this method along with the
`manualStatus`. The first returned boolean is true if the system should
be seen as ready and the second is true if the override does handle the
given request. If it does not, then we will fall back to our normal
is-ready check.