status: remove err when marking sub-server as running
Ensure that the sub-server error string is set to an empty string when
marking a sub-server as running. This will be important for the next
commit, which will retry to start the lnd sub-server even if we error
when starting the lnd-client(s), and therefore may mark the lnd
sub-server as running even after we've errored when starting the
lnd-client(s).