README: clarify the Compatibility section
The wording for the first table of the Compatibility section of the
README was confusing, as it didn't clearly state that the table showed
the _minimum_ `lnd` version that can be used for different litd
versions, when lnd is running in remote mode.

This commit updates the wording in that section to make that clearer.