firewalldb: fetch correct bucket for session level temp KV store
In this commit, we fix a bug which would result in the "perm" store
being used for the LocalTemp KV DB _if and only if_ the KV store being
requested was a session level one (as opposed to a session feature level
one). It is ok for us to just change this here since as of this commit,
the session level KV store is not used at all and so nothing will be in
that store yet.