mod: bump lnd, Go, lndclient, loop, pool and tapd
Bump lnd to `v0.18.0-beta`. This also bumps `lndclient` to `v0.18.0-1`
and loop to `v0.28.5-beta`. Note that since loop `v0.28.5-beta` depends
on the lnd  `v0.18.0-beta.1` tag, litd will also use that tag. That tag
is still referencing the same commit as lnd `v0.18.0-beta` though, so
that's not an issue.

We also need to bump the `pool` and
`taproot-assets` daemons to branch specific versions that support
lnd `v0.18.0-beta`. Finally, the `pool` auctioneerrpc package is also
bumped to `v1.1.2`.

We also adapt it itests for the new lnd version.

Finally as lnd `v0.18.0-beta` uses go 1.22.3, we also bump the Go
version in the go.mod file.