mod+itest: reproduce and fix reverse direction invoice payment
Fixes a bug that when an invoice is paid, the full invoice amount is
checked when validating the local channel balance and reserve. But for
custom channels the actual HTLC amount might be just above dust and
shouldn't validate the channel reserve.
We update our itest to reproduce the issue and then update to the lnd
commit that fixes it.