itest: make shard size configurable
We don't need to set the shard size in every test, as the max_htlc value in the actual routing network will force splits if needed. When we go to send a large payment, with this value, and the default max parts value of 16, we'll never be able to send enough shards to finalize a payment amount.