mod+itest: assert co-op close when initiator has zero balance
This PR bumps the version of tapd to fix the bug of panicking when an
HTLC that pays the full balance to the recipient of a channel is being
created.
We assert this no longer is a problem and we also assert that the final
balance when co-op closing in that state (where the initiator has zero
balance) is correct.