config+itest: move custom message override
It turns out the custom message override that should've allowed tapd to
send custom p2p wire messages never worked, because we set the value
after calling into lnd.ValidateConfig() which actually injects the value
into the lnwire package. We now move the code and also create an itest
for it (we should be able to try funding multiple times after each
other, as each attempt should be cleaned up by the receiver when they
receive the error message).