litcli: relax fundchannel checks
The RPC calls were updated to allow composite UTXOs (assets from
multiple asset utxos) for funding capability, but the same functionality
is not available in the CLI due to the existing sanity check.

This commits changes the sanity check to allow for composite UTXOs by
validating whether the running sum of the assets in the UTXOs is greater
than the amount to be funded.