itest: update WaitUntilStarted to only use status server
This is so that we can use this Wait function to check the state of
sub-servers even when LiT is running in stateless init mode. As
currently, a macaroon is used to make a call to each sub-server which we
wont have access to in stateless init mode. This commit also adds a new
WaitForLNDWalletReady call which only waits for LND's wallet to be
ready. We'll use this during the set-up of a stateless init node in an
itest so that we can know when we are able to Create the wallet.