Revert "make: update VERSION_TAG to a globally-consistent build-input identifier"
This reverts commit 66d2bd1bdade3cacf1df263799bf4569a97af3ec.
Unfortunately, this commit leads to a "-dirty" suffix being added to
built releases, on multiple environments. The reason for that is that as
the "make release" build process rebuilds the app build folder, which
temporarily removes the .gitkeep file in that folder. As the VERSION_TAG
then gets based on a build that lacks the .gitkeep file, the build
process adds the "-dirty" suffix to the version.