multi: bump `loop` and `swapserverrpc`
This commit updates the `loop` dependency to version `v0.28.8-beta`.
Notably, the actual `loop` reference points to a branch based on top of
the `v0.28.8-beta` tag, which modifies `v0.28.8-beta` to use `lnd`
version `v0.18.3-beta` rather than `v0.18.3-beta.rc1`.

Additionally, we upgrade `github.com/lightninglabs/loop/swapserverrpc`
to version `v1.0.10`.
As `loop/swapserverrpc` `v1.0.10` references the `reservation.proto`, we
need to modify the `build-protos.js` script to include it as on of the
fetched protos files.