itest: expect failure on direct rfq peer btc invoice
Previously we'd consider it acceptable to settle direct rfq peer
invoices, which included no rfq scid, with asset HTLCs. This behavior
has been updated on the tapd invoice manager and we no longer accept
asset HTLCs on invoices that do not expect assets. This commit updates
such payments in our itests to instead expect a failure.