accounts: return consistent error for duplicate label
In preparation for when we have a SQL DB implementation, we want our
unit tests to run smoothly against all DB backends and have the same
results. To achieve this, we need to turn some errors into global error
variables that can be matched against instead.

In this commit, we do this for the unique constraint violation of the
account label.