multi: add CreditAccount & DebitAccount endpoints
This commit introduces the structure for new endpoints in the accounts
subsystem, enabling balance adjustments by crediting or debiting a
specified amount. This contrasts with the existing `UpdateAccount`
implementation, which directly sets the balance to a fixed value.

For more details on the drawbacks of the current implementation, see
[issue #648](https://github.com/lightninglabs/lightning-terminal/issues/648).

The actual implementation of the endpoints will be introduced in
subsequent commits.