itest: use groupkey on grouped asset & liquidity edge cases itests
We add support for group keys on payments and invoices across the
itests. This includes all the required helpers and optional arguments to
make adding invoices and sending payments group-key enabled.

In order to run the liquidity edge cases twice, once using only asset
IDs and once using only group keys, we need to extract the main logic
into a re-usable function that has the group key option on the top
level. If the group key flag is set, then we append the respective group
key opt to the payment & invoice related calls.