make: compile binaries in itest-only
The itest-only goal is really useful for not needing to run the app
build. But when switching between lnd versions, it's annoying if the lnd
binary isn't built with itest-only. So we change that and always also
build the btcd/lnd binaries with itest-only. This will take slightly
longer but lead to fewer issues when switching branches between
different lnd major versions.