firewalldb: kvdb test file
In preparation for testing each of the existing unit tests against
various types of DB backends (ie, different implementations of the
RulesDB interface), we clean-up the test code such that the RulesDB is
always constructed through the same `NewTestDB` function. Eventually, we
will add different implementations of this function which will build
under different build flags.