itest: add groupkey mode to liquidity edge cases
In order to run the liquidity edge cases twice, once using only asset
IDs and once using only group keys, we need to extract the main logic
into a re-usable function that has the group key option on the top
level. If the group key flag is set, then we append the respective group
key opt to the payment & invoice related calls.