itest: use Charlie as proof courier, remove Zane
We remove the Zane node from all tests where its only role was being the
universe server. Starting a node takes multiple seconds, so if we can
spin up fewer nodes, we can save some time.
And functionality wise Charlie can easily be the universe, we just need
to define its port upfront so we can configure it to be its own proof
courier.
The topology for all tests is at least the following:

Charlie  --[assets]-->  Dave  --[sats]-->  Erin  --[assets]-->  Fabia

Which means, we the case where one of the channel peers is its own
universe (Charlie in Charlie<->Dave) as well as the case where the
universe server is a remote node (Charlie in the Erin<->Fabia channel).