firewalldb: move generic DB setup code to the DB.Start method
In this commit, a new `DeleteTempKVStores` method is added to the
`RulesDB` interface and this is called in the `DB`'s `Start` method.
This is done because regardless of the backend that we use for the
RulesDB, we want this clean-up code to run at start time. It also makes
more sense to have this call in a Start method rather than in a
constructor.