firewalldb: assert strong coupling between kvstores and sessions
If the is writing to a kvstore that is namespaced by a session's ID, we
first assert that the session exists in the session store. We add a test
for this accordingly. We do this in preparation for adding a sql backend
which will error out if we try to link a kvstore to a session that does
not yet exist.