itest: add safe HTLC failure edge case for custom channels
We enhance the liquidity edge cases itest with a new case which makes a
payment attempt over a last-hop that uses assets, and does not have
sufficient asset liquidity. The healthy expectation is for the HTLCs to
be added, as there exists partial liquidity, but they should eventually
fail. As a sanity check we then end things with a small payment that can
succeed, and assert its completion.