multi: update `GetInfoResponse` response
Add the `tag` and `commit_hash` fields to the GetInfoResponse. The
semantics of the `version` field is also updated to always contain the
most recent semantic version of the litd node, following the semantic
versioning 2.0.0 spec (http://semver.org/).

This commit also updates the `litcli --version` command result, to
include all three fields of the GetInfoResponse.